Move code from sample to PayEx.Client
