+++ src/PayEx.Client/ISelectClient.cs
@@ -1,0 +1,7 @@
+namespace PayEx.Client
+{
+    public interface ISelectClient
+    {
+        string Select();
+    }
+}
