+++ src/PayEx.Client/PayExHttpClient.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Text;
@@ -86,7 +88,10 @@ namespace PayEx.Client
             }
             else
             {
-                problems = new ProblemsContainer("Other", $"Response when calling PayEx was: '{response.StatusCode}'");
+                IEnumerable<string> customHeader = null;
+                _client.DefaultRequestHeaders.TryGetValues("X-Payex-ClientName", out customHeader);
+                var aggr = customHeader != null? customHeader.Aggregate((x, y) => x + "," + y) : "no-name";
+                problems = new ProblemsContainer("Other", $"Response when calling PayEx `{aggr}` was: '{response.StatusCode}'");
             }
 
             var ex = onError(problems);
